feat: separate fixtures
